+++ client/features/media/components/ImagePreview.tsx
@@ -159,7 +159,7 @@ const createStyles = (theme: Theme, topInset: number) =>
       height: "100%",
     },
     loadingOverlay: {
-      ...StyleSheet.absoluteFillObject,
+      ...StyleSheet.absoluteFill,
       justifyContent: "center",
       alignItems: "center",
       backgroundColor: "black",
@@ -171,7 +171,7 @@ const createStyles = (theme: Theme, topInset: number) =>
       fontSize: 14,
     },
     errorOverlay: {
-      ...StyleSheet.absoluteFillObject,
+      ...StyleSheet.absoluteFill,
       justifyContent: "center",
       alignItems: "center",
       backgroundColor: "black",